Train the keras model
Input: time
Output: true anomaly
Note that we are training on a single orbit. If we were training on multiple orbits, we would also need initial state as an input. However, orbital elements other than true anomaly are constant for a single orbit propagation, so that s not necessary here.